autocc.py: Fix generated port lists that end with comment
For the last port of a DUT, autocc currently appends `_2,` at the end of
the line. However, if the line ends on a comment, the `_2,` suffix is
appended *behind* the comment instead of the variable name, causing
syntax errors. In particular, this affects port lists ending with an
output since autocc appends `\\output` after those. Fix this by
inserting the suffix in lines that contain a comment but no comma (last
port) *before* the comment.

Signed-off-by: Nils Wistoff <nwistoff@iis.ee.ethz.ch>